feat(mcp): make the MCP tools + LSP web-compatible
package:apollovm/apollovm_mcp.dart is now free of dart:io and dart:isolate, so
the MCP server and every apollovm.* / apollovm.lsp.* tool compile and run in a
browser (dart2js/DDC). Construct ApolloMcpServer on any StreamChannel<String>
and drive the tools in-process.

- Split the dart:io transports + CLI (serveStdio, HttpSseTransport, CommandMcp)
  into a new native-only library package:apollovm/apollovm_mcp_io.dart, which
  re-exports apollovm_mcp.dart. bin/ and the transport/CLI tests import it.
- Platform-select the per-tool timeout executor (like wasm_runtime.dart):
  killable-isolate on native, in-process cooperative-timeout on web. New
  isolate_executor.dart facade + _io (dart:isolate) + _generic (web) variants.
- Add cross-platform test/mcp/web_compat_test.dart (runs under
  dart test --platform chrome) guarding the web-safe surface. Verified: dart2js
  compiles the surface; 6 web-compat tests pass in Chrome; all 75 MCP tests
  pass on the VM.
