+++ test/mcp/http_sse_transport_test.dart
@@ -6,7 +6,7 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:apollovm/apollovm_mcp.dart';
+import 'package:apollovm/apollovm_mcp_io.dart';
 import 'package:test/test.dart';
 
 void main() {
